Misc changes
A) Run ossdevlinks at soundoff to save legacy_devices then. Save /dev/dsp
   link as well. (This helps people who change default symlink not have it reset
   at boot).
B) Make ossxmix a bit more resilient to USB device removals.
C) setup/FreeBSD/build.sh should create /usr/lib/oss/include/internals on
   FreeBSD as well.
D) srcconf.c: Use "-O3" instead of "-O6". It's the same for gcc, but
   clang/llvm complains. Also add a do { } while(0) loop to
   kernel/framework/include/oss_config.h to avoid clang warning.

+++ setup/Linux/build.sh
@@ -105,7 +105,11 @@ cp -a $SRCDIR/setup/Linux/oss/* prototype/$OSSLIBDIR/
 cp -a $SRCDIR/setup/Linux/sbin prototype/usr/
 chmod +x prototype/$OSSLIBDIR/scripts/*
 
-ld -r -o prototype/$OSSLIBDIR/$OBJECTS/osscore.o target/objects/*.o
+if ! ld -r -o prototype/$OSSLIBDIR/$OBJECTS/osscore.o target/objects/*.o
+then
+  echo Linking osscore failed!
+  exit 1
+fi
 
 rm -f devlist.txt devices.list
